implementation/fastapi-rbac/requirements.txt
# FastAPI RBAC Requirements

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Caching
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# Type checking
mypy>=1.8.0
